refactor: move new assignee logic to separate file
